memperbaiki ui dan mengintergrasi database dengan ui agar ui terlihat lebih dinamis dan berusaha menyamakan fungsi sesuai yg di minta pihak jcc

+++ database/migrations/2021_10_08_173527_create_master_barang_table.php
@@ -16,7 +16,7 @@ class CreateMasterBarangTable extends Migration
         Schema::create('master_barang', function (Blueprint $table) {
             $table->id();
             $table->string('nama_barang');
-            $table->integer('harga_satuan');
+            $table->float('harga_satuan');
             $table->timestamps();
         });
     }
